Change timer to be more like a timer
Looking at java.util.Timer, SysTick Timer, and JavaScript Timers one
can see that a timer is meant for one purpose and that is to do
something when a time has elapsed. All of the other controls do not
make much sense.

The interface has been reduced down to 3 core functions:

- is_running()
- clear()
- schedule()

Resolves #44