can::attach_interrupt handler takes can::messages
Taking a reference to a can object was a design flaw.

New behavior for canbus driver implementations are:

- When a message is recieved, the canbus driver must copy that message
  to a embed::can::message object, clear the interrupt associated with
  that can message, then pass the message to the handler.

The biggest issue with the previous post is that it pushed the
responisiblity of clearing the interrupts on the user passing the
handler to the attach_interrupt() function. A user may not understand
that in order to clear the interrupt they need to run receive() to copy
the stored message and clear any associated interrupt flags. This would
result in their handler being called repeatedly. The new design pushes
the clearing action on the driver where it should be.