Remove reception functions from embed::can
The previous function signature for attach_interrupt() begs the
question of “who clears the interrupt handler and when”? On many
systems the way it is cleared is either to read the data out of
the FIFO/mailbox/registers or set a bit that clears the FIFOs.

This leaves a few options:

- User defined handler are REQUIRED to read the buffers or be stuck
  in a loop.
- When a handler occurs, user defined handler is called, when finished
  service routine will clear all messages in the buffer.
- Pass messages to handler, clear message after usage.

The last choice is the preferred option because the handler has no
obligation to read the buffer or even access the can object and
clearing of each message can be handled by the service routine.

The reception functions were removed because they no longer useful
with this new API since has_data() and reception will always return
empty if an interrupt services these first.