Rename driver::reset() to set_as_uninitialized()
Reset is a common term used in many APIs and having driver take
the name seems like a poor decision especially when the name doesn't
make too much sense in terms of what its meant to do. Where as
"set_as_uninitialized()" directly states exactly the behavior that is
occuring with the driver. It also makes is clear that it doesn't do any
work but merely marks the driver as uinitialized.