Add uptime counter class
uptime counter class is a solution to the uptime calculation problem
that counter and frequency have. The issue has to do with frequency's
duration_from_cyles() function. It takes a 64-bit number and performs
multiplication against a 32-bit number. The resultant for such a number
would need to be 96 bits in width. Once the input count reaches
a specific size, the result can no longer fit within a 64-bit number,
the result is invalid. The resulting std::chrono::nanoseconds object's
count() value, will still be far away from the 64-bit maximum.

To solve this, uptime counter does not take the whole uptime but
calculates the delta between the previous time uptime was called and the
current time. Then uses the count delta to calculate the nanoseconds
delta between uptime calls and adds that to a sum of nanoseconds since
the start of counting. This class will work so long as the delta
between uptime calls does not exceed a specific time duration.
This comes out to be 15 weeks.