Use quanity types rather than unit types
unit types are symbols with no data but compile time information about
what a unit is. A quanity type is magnetude of a unit type. In order to
use actual numeric quantities in the code, the units need to be used in
quantity types. This change adds quanity types for acceleration and
temperature.

disabled error.test.hpp and made visibility hidden for each test
translation unit to prevent the tests from seg faulting and to remove
the weak reference warnings, respectively.