Add config::float_t
The application user can use this configuration option to change the
default floating point type of the whole system.

This allows developer to choose if they want the system to default to
32-bit single precision floating point types or 64-bit double precision
floating point types. Developers may want to choose "float" as their
float type if they have a hardware FPU. If an application is run on a
target with a double precision FPU then this flag can be set to double.

Resolves #340