Doxygen: Use EXCLUDE_PATTERNS
Previously EXCLUDE directive was used to exclude the "third_party"
directory. This required a path that included the project name within
it. This results in a break if the directory structure ever changes
like it did when the project name was changed to libhal.

Rather that using a direct link EXCLUDE_PATTERN can be used instead to
blanket exclude any "third_party" directory under the include/
directory.