Add units as config::float aliases
Migrating the codebase to simple use floats (or percentage where
applicable) for units. This would mean that there will be no compile
time checks or conversion when it comes to units. The units are merely
annotations to help better document what parameters are what, but
underneath they are simply floats.

This way we can see how users use these units and what patterns can be
introduced to make conversion easier or disallow certain catastrophic
errors. Going forward, these types, if they were to evolve, will need to
be implicitly convertable to floats as to not break future code, which
introduces safety concerns, but seems to be the middle ground between
using plain old floats and strongly typed units which made working with
them burdensome, cumbersome, confusing, and slow.

Issue #230