Remove templates from interfaces
The added templates create space for added complexity where it wasn't
before for a very small subset of situations. It is better to have the
base interfaces be simple and support the majority of use cases and
provide more advanced interfaces for more complicated or specific use
cases. Most ADCs are 24 bits and below in precision. A double floating
point ADC percentage is only necessary above that range and thus, if it
is needed a `hal::precision_adc` can be introduced. The same goes for
all other interfaces.

- Rename hal::percentage to hal::percentage_t
- hal::percentage is an alias for type percentage<float>