Replace `hal::counter` -> `hal::steady_clock`
`hal::steady_clock` represents a monotonically incrementing steady time
keeping mechanism. The interval between ticks of the clock are
consistent and the uptime of a steady clock never decreases.

`hal::counter` has an interface that gave the impression that the
frequency of the counter could change, which can result in confusing and
inconsistent behavior if true. A steady clock with hard requirements for
how it must operate can side step these issues entirely.

The name counter is overloaded with a ton of non-time based thing so a
name relating to time, clocks, made more sense.

Resolves #415