hal::serial combine read + bytes_available
Rather than make these two seperate calls, in general the information
about the number of bytes remaining and the capacity is useful after
each read operation. This is due to the idea it is unlikely that a
user will ever run bytes_available and not also then run read. In
general drivers that read from serial simply want to read whats
available from the serial working buffer.

write now returns a "write_t" which only includes a size_t variable for
now. This will allow for further expansion in the future if it is deemed
necessary.

Issues #293