Add hal::as_bytes() hal::as_writable_bytes
`hal::as_bytes()` converts any container, C-style array or pointer and
size set to a span of bytes. This is similiar to std::as_bytes() except
that function only works on spans to be begin with. This works with any
structure with a `data()` and `size()` class function as well as C-style
arrays.

Bump boost-leaf version to work with latest on remote.

Resolves #452