Remove boost::ut::throws calls w/ Boost.LEAF
A call to boost::leaf::result::value() when there is no value executes
an assert and this will abort the program at that location rather than
throwing. The behavior doesn't seem to be consistent across platforms,
but on M1 mac systems, it aborts in the middle of the unit test runtime
execution.

Removing this will make unit tests workable on M1 machines and more
consistent everywhere else.