Use iosfwd correctly for can::message_t
The stream operator must be (1) a template and (2) the generic
basic_ostream type.

Templates will prevent the compiler from complaining if the forwarded
types are not defined so long as the template function isn't used.

Using a basic_ostream type allows other ostream variants to be used
rather than the built in std::ostream.