Add hal::skip_past(serial) functor
Skips bytes receieved on the serial port until a specified sequence
is found. It is non-blocking in that it will not wait for data to come
in there isn't any available. It also has a read limit, to adjust how
many bytes of data to read before giving control back to the caller.

With each call to the functor, it will parse more data out of the serial
port until it finds the sequence. At which point it will no longer read
data from the serial port and will always return finished.

- Bumb version of clang-format to 15
- Add work_state enumeration to enum.hpp
- Bumb version to 0.1.18
- Cleanup callback recorder parameters

Resolves #526