Fix boost leaf macro issues in tweaks file
The `error.hpp` file is main include that brings in Boost.LEAF as well
as the libhal.tweaks file via the config include. The position of the
config include is below Boost.LEAF which results in the macro
definitions being ignored. Moving the "config.hpp" to the top of the
file will change this.

Resolves #561