:heavy_plus_sign: Add & use tl::function_ref in all APIs
tl::function_ref is an implementation of the future std::function_ref.
When std::function_ref is released, `hal::function_ref` will point to it
instead of `tl::function_ref`.

Migrate from `std::function` to `hal::function_ref` because the former
is owning, potentially allocating, and results in copies, where as
the later is non-owning, never allocates and is the size of just two
pointers.

Bump version to 0.3.1