:zap: Improve `pwm` interface performance & docs
`hal::pwm` now has a full explanation regarding what are the possible
input values `pwm::duty_cycle` can take. The performance boost comes
from putting the clamp function into the interface, the first of such
situations in libhal, in order to create a guarantee about the value
passed to the driver. This guarantee allows callers and implementors
to eliminate checks and clamping code and manage all of this in one
central place, reducing code bloat.