:rocket: Launch version 1.0.0
libhal is a dependency for the SJSU Robotics team and thus its SemVer
number must be set to 1.0.0. This is important because it will help with
communicating to our users when APIs break, when features are added
and when patches are added. This isn't possible with versions below
1.0.0 as stated in SemVer.

The following changes were made to improve flexibility, improve
extensibility and remove some technical debt.

- :zap: Add input clamp for `hal::dac` write operation
- :zap: Add input clamp for `hal::motor` power operation
- :memo: Update documentation for each interface and each interface
  building block
- :boom: All non configuration and non interrupt APIs return structures

Resolves #477
Resolves #652