:fire: Remove socket interface for 2.0.0
The socket interface was a rushed construction with some problematic
issues. For now, its best to not include it and consider what the
network stack should be like for libhal in the future and if a socket
interface is even warranted.

This change will overwrite 2.0.0 version. Nothing that currently uses
2.0.0 also uses sockets so this should be fine for now.